Extract inventory item repository
