+++ lib/business_logic/item_action.dart
@@ -18,6 +18,8 @@ import 'package:spare_parts/widgets/print_dialog/print_dialog_mobile.dart'
     if (dart.library.html) 'package:spare_parts/widgets/print_dialog/print_dialog_web.dart';
 import 'package:spare_parts/widgets/user_avatar.dart';
 
+import '../services/repositories/repositories.dart';
+
 enum ItemActionType { delete, edit, borrow, release, assign, print }
 
 abstract class ItemAction {
@@ -63,10 +65,13 @@ class DeleteItemAction extends ItemAction {
 
   @override
   handle(BuildContext context, InventoryItem item) {
-    final firestoreService = context.read<FirestoreService>();
+    final inventoryItemRepository = context.read<InventoryItemRepository>();
 
     commonHandle(
-      () => firestoreService.deleteItem(item.id),
+      () async {
+        await inventoryItemRepository.delete(item.id);
+        return true;
+      },
       context,
       'deleted',
     );
@@ -85,7 +90,7 @@ class AssignItemAction extends ItemAction {
 
   @override
   handle(BuildContext context, InventoryItem item) {
-    final firestoreService = context.read<FirestoreService>();
+    final inventoryItemRepository = context.read<InventoryItemRepository>();
     final callableService = context.read<CallableService>();
 
     commonHandle(
@@ -116,7 +121,7 @@ class AssignItemAction extends ItemAction {
         item.borrower = userIds.isEmpty
             ? null
             : users.firstWhere((u) => u.id == userIds.first).toCustomUser();
-        await firestoreService.updateItem(item.id, item);
+        await inventoryItemRepository.update(item.id, item);
         return true;
       },
       context,
@@ -138,6 +143,7 @@ class BorrowItemAction extends ItemAction {
   @override
   handle(BuildContext context, InventoryItem item) {
     final firestoreService = context.read<FirestoreService>();
+    final inventoryItemRepository = context.read<InventoryItemRepository>();
     final auth = context.read<FirebaseAuth>();
 
     commonHandle(
@@ -175,7 +181,7 @@ class BorrowItemAction extends ItemAction {
           }
         }
 
-        await firestoreService.borrowItem(
+        await inventoryItemRepository.borrow(
           item,
           CustomUser.fromUser(user),
         );
@@ -209,6 +215,7 @@ class ReleaseItemAction extends ItemAction {
   @override
   handle(BuildContext context, InventoryItem item) {
     final firestoreService = context.read<FirestoreService>();
+    final inventoryItemRepository = context.read<InventoryItemRepository>();
     final auth = context.read<FirebaseAuth>();
 
     commonHandle(
@@ -220,7 +227,7 @@ class ReleaseItemAction extends ItemAction {
           createdAt: DateTime.now(),
         );
         await firestoreService.addEvent(item.id, event);
-        await firestoreService.releaseItem(item);
+        await inventoryItemRepository.release(item);
 
         return true;
       },
