+++ lib/pages/home_page/borrowed_items_view.dart
@@ -2,22 +2,23 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:spare_parts/entities/inventory_item.dart';
-import 'package:spare_parts/services/firestore_service.dart';
 import 'package:spare_parts/widgets/empty_list_state.dart';
 import 'package:spare_parts/widgets/error_container.dart';
 import 'package:spare_parts/widgets/inventory_list_item.dart';
+
+import '../../services/repositories/repositories.dart';
 
 class BorrowedItemsView extends StatelessWidget {
   const BorrowedItemsView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final firestoreService = context.read<FirestoreService>();
+    final inventoryItemRepository = context.read<InventoryItemRepository>();
     final auth = context.read<FirebaseAuth>();
 
     return Center(
       child: StreamBuilder<List<InventoryItem>>(
-        stream: firestoreService.getItemsStream(
+        stream: inventoryItemRepository.getItemsStream(
           whereBorrowerIs: auth.currentUser?.uid,
         ),
         builder: (context, snapshot) {
