+++ lib/services/repositories/repositories.dart
@@ -1,0 +1,1 @@
+export 'inventory_item_repository.dart';
